[Desktop Zoom] Update default fieldtrial experiment
This CL updated the fieltrial config for the desktop zoom feature to use
a non-default value, which is the eventual launch candidate. The fact
that we are scaling content 9% in some cases is not meaningful to
include in render tests, and so we disable the feature in some cases.

We update a rounding choice for the HostZoomMap as it is not necessary
and allows us to exactly match values with floating point numbers across
code boundaries, which requires us to update some test values. We
disable tests for ContextualSearch and ContextMenu, as well as wpt
tests, since their pixel-based clicks will require updating for Desktop
once the feature is fully landed at a chosen zoom value (all these tests
pass on each device except android-desktop).

Bug: 450281745
Include-Ci-Only-Tests: chromium.android.desktop:android-desktop-x64-rel-15-tests|android_chrome_wpt_tests,chrome_public_test_apk
Validate-Test-Flakiness: skip
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7522308
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1578313}

+++ chrome/android/java/src/org/chromium/chrome/browser/app/ChromeActivity.java
@@ -2067,10 +2067,12 @@ public abstract class ChromeActivity extends AsyncInitializationActivity
             if (!DisplayUtil.isContextInDefaultDisplay(this)
                     && DeviceFormFactor.isNonMultiDisplayContextOnTablet(this)) {
                 HostZoomMap.setTransparentZoomAdjustment(
-                        ContentFeatureList.sAndroidMonitorZoomScalingFactor.getValue() / 100.0f);
+                        (float) ContentFeatureList.sAndroidMonitorZoomScalingFactor.getValue()
+                                / 100.0f);
             } else if (DeviceInfo.isDesktop()) {
                 HostZoomMap.setTransparentZoomAdjustment(
-                        ContentFeatureList.sAndroidDesktopZoomScalingFactor.getValue() / 100.0f);
+                        (float) ContentFeatureList.sAndroidDesktopZoomScalingFactor.getValue()
+                                / 100.0f);
             }
         }
     }
